Load Previous Work [1st Major Commit]
I’ve previously set up Git, but am unsure of where I left it. Instead
of spending the time to reacquaint myself, I’m going to commit a full
working version of Zurb’s foundation with some basics set up.

I am still not sure if this is a proper Ruby on Rails setup or not.